Split out `RasterFunctions` into separate traits based on function type for easier navigation and maintenance.
